removes parallel routing for a layout
