--- web/app/chat/[[...conversation_id]]/@chatList/page.tsx
@@ -1,9 +1,0 @@
-import ChatList from '@/components/chat/ChatList';
-
-export default async function Index() {
-  return (
-    <>
-      <ChatList />
-    </>
-  );
-}
