+++ web/app/chat/page.tsx
@@ -1,0 +1,17 @@
+import Chat from '@/components/chat/Chat';
+import { createClient } from '@/utils/supabase/server';
+import { cookies } from 'next/headers';
+
+export default async function Index() {
+  const cookieStore = cookies();
+  const supabase = createClient(cookieStore);
+  const {
+    data: { user },
+  } = await supabase.auth.getUser();
+
+  return (
+    <>
+      <Chat conversation={null} currentUser={user} />
+    </>
+  );
+}
